+++ index.html
@@ -9,250 +9,9 @@
       rel="stylesheet"
       href="https://cdn.jsdelivr.net/npm/reset-css@5.0.1/reset.min.css"
     />
-    <style>
-      /* 공통 */
-      h1 {
-        font-weight: 800;
-        font-size: 24px;
-        line-height: 36px;
-      }
-
-      h2 {
-        font-weight: 600;
-        font-size: 20px;
-        line-height: 24px;
-        initial-letter: 0.15px;
-      }
-
-      .button--primary {
-        background-color: #4e5ba6;
-        color: #ffffff;
-        padding: 10px 12px;
-        border: 0px;
-        border-radius: 4px;
-        font-weight: 700;
-        font-size: 14px;
-        line-height: 16px;
-        letter-spacing: 1.25px;
-      }
-
-      .input {
-        height: 36px;
-        border: 1px solid #b4b4b4;
-        border-radius: 4px;
-        padding: 0;
-      }
-
-      /* 헤더 */
-      header {
-        height: 64px;
-        background-color: #4e5ba6;
-        color: #ffffff;
-        display: flex;
-        align-items: center;
-        padding-left: 64px;
-      }
-
-      /* 푸터 */
-      footer {
-        height: 80px;
-        border-top: 1px solid rgba(0, 0, 0, 0.12);
-        color: #4e5ba6;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-      }
-
-      /* 메인 */
-      main {
-        margin: 80px 0;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-      }
-
-      main > article {
-        border: 1px solid rgba(0, 0, 0, 0.12);
-        border-radius: 4px;
-        width: 440px;
-        height: 718px;
-        padding: 36px 16px;
-        box-sizing: border-box;
-        display: flex;
-        flex-direction: column;
-        gap: 28px;
-        font-weight: 400;
-        font-size: 15px;
-        line-height: 24px;
-        letter-spacing: 0.5px;
-      }
-
-      /* 제목 */
-      .lotto-game__title {
-        align-self: center;
-      }
-
-      /* 구입금액 */
-      .purchase-amount__content {
-        margin-top: 4px;
-        display: flex;
-        justify-content: space-between;
-        gap: 16px;
-      }
-
-      .purchase-amount__content__input {
-        flex-grow: 1;
-      }
-
-      .purchase-amount__content__input {
-        padding-left: 12px;
-      }
-
-      /* 구매한 로또 */
-      .purchased-lottoes {
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-      }
-
-      .purchased-lottoes__content {
-        height: 260px;
-        overflow-y: auto;
-        display: flex;
-        flex-direction: column;
-        gap: 4px;
-      }
-
-      .purchased-lottoes__content__item {
-        display: flex;
-        align-items: center;
-        padding: 4px 0;
-        gap: 4px;
-      }
-
-      .purchased-lottoes__content__item__emoji {
-        font-size: 24px;
-      }
-
-      /* 당첨번호, 보너스 번호 */
-      .winning-lotto {
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-      }
-
-      .winning-lotto__content {
-        display: flex;
-        justify-content: space-between;
-      }
-
-      .winning-lotto__content__winning-numbers {
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-      }
-
-      .winning-lotto__content__bonus-number {
-        display: flex;
-        flex-direction: column;
-        align-items: end;
-        gap: 8px;
-      }
-
-      .winning-lotto__input {
-        width: 34px;
-        text-align: center;
-      }
-
-      /* 모달 */
-      .modal-wrapper {
-        position: fixed;
-        top: 0;
-        left: 0;
-        width: 100%;
-        height: 100vh;
-        display: flex;
-        justify-content: center;
-        align-items: center;
-        background-color: rgba(0, 0, 0, 0.5);
-      }
-
-      .modal {
-        background-color: #ffffff;
-        border-radius: 4px;
-        width: 328px;
-        height: 440px;
-        padding: 16px 16px 28px;
-        display: flex;
-        flex-direction: column;
-      }
-
-      .modal__button--close {
-        align-self: flex-end;
-        background-color: #ffffff;
-        border: none;
-        font-weight: 700;
-        font-size: 20px;
-        line-height: 16px;
-        letter-spacing: 1.25px;
-      }
-
-      .modal__body {
-        flex-grow: 1;
-        display: flex;
-        flex-direction: column;
-        justify-content: space-between;
-        align-items: center;
-      }
-
-      .modal__body__content {
-        width: 100%;
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        gap: 36px;
-      }
-
-      .modal__body__content__table {
-        display: table;
-        width: 100%;
-        border-collapse: collapse;
-      }
-
-      .modal__body__content__table__row {
-        display: table-row;
-        border-top: 1px solid;
-        border-bottom: 1px solid;
-        font-weight: 400;
-        line-height: 36px;
-        font-size: 15px;
-        letter-spacing: 0.5px;
-      }
-
-      .modal__body__content__table__row:first-child {
-        font-weight: 600;
-      }
-
-      .modal__body__content__table__cell {
-        display: table-cell;
-        text-align: center;
-      }
-
-      .modal__body__content__profit-rate {
-        font-weight: 700;
-        font-size: 15px;
-        line-height: 24px;
-        letter-spacing: 0.5px;
-      }
-
-      .modal__body__button {
-        align-self: stretch;
-      }
-    </style>
   </head>
 
   <body>
     <div id="app"></div>
-    <script type="module" src="./src/js/index.js"></script>
   </body>
 </html>
