+++ src/view/step2/components/PurchaseAmount/index.js
@@ -1,3 +1,5 @@
+import "./style.css";
+
 class PurchaseAmount {
   #$parent;
   #props;
