+++ src/view/step2/components/PurchasedLottoes/index.js
@@ -1,3 +1,5 @@
+import "./style.css";
+
 class PurchasedLottoes {
   #$parent;
   #props;
