+++ src/view/step2/components/ResultModal/style.css
@@ -1,0 +1,84 @@
+/* 모달 */
+.modal-wrapper {
+  position: fixed;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  background-color: rgba(0, 0, 0, 0.5);
+}
+
+.modal {
+  background-color: #ffffff;
+  border-radius: 4px;
+  width: 328px;
+  height: 440px;
+  padding: 16px 16px 28px;
+  display: flex;
+  flex-direction: column;
+}
+
+.modal__button--close {
+  align-self: flex-end;
+  background-color: #ffffff;
+  border: none;
+  font-weight: 700;
+  font-size: 20px;
+  line-height: 16px;
+  letter-spacing: 1.25px;
+}
+
+.modal__body {
+  flex-grow: 1;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.modal__body__content {
+  width: 100%;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 36px;
+}
+
+.modal__body__content__table {
+  display: table;
+  width: 100%;
+  border-collapse: collapse;
+}
+
+.modal__body__content__table__row {
+  display: table-row;
+  border-top: 1px solid;
+  border-bottom: 1px solid;
+  font-weight: 400;
+  line-height: 36px;
+  font-size: 15px;
+  letter-spacing: 0.5px;
+}
+
+.modal__body__content__table__row:first-child {
+  font-weight: 600;
+}
+
+.modal__body__content__table__cell {
+  display: table-cell;
+  text-align: center;
+}
+
+.modal__body__content__profit-rate {
+  font-weight: 700;
+  font-size: 15px;
+  line-height: 24px;
+  letter-spacing: 0.5px;
+}
+
+.modal__body__button {
+  align-self: stretch;
+}
