+++ src/view/step2/components/WinningLotto/index.js
@@ -1,3 +1,4 @@
+import "./style.css";
 import WinningNumbers from "./WinningNumbers";
 import BonusNumber from "./BonusNumber";
